backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.28.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
httpx==0.27.2
ldap3==2.9.1
PyJWT==2.10.1
msal==1.31.1
apscheduler==3.10.4
python-dateutil==2.9.0
